+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
-### Next release
+### 1.44.0
+
+### New features
 
 - Added `Wizaplace\SDK\Vendor\Order\Payment::getCommitmentDate`
-- Fixed `Wizaplace\SDK\User\UserService::enable` and `Wizaplace\SDK\User\UserService::disable`
+
+### Corrections
+
+- Fixed an issue with `Wizaplace\SDK\User\UserService::enable` and `Wizaplace\SDK\User\UserService::disable`
 
 ## 1.43.0